feat(routing): add routing
- create routes object with routesConfig.tsx
- create RouteProvider in index.tsx
- create page folder for gathering views
- change App file to contain only routes of the App
- move all previous App component to Home page component without changes
- move tApp.test.tsx to Home.test.tsx without changes
