+++ src/routes/index.tsx
@@ -1,0 +1,8 @@
+import { RouterProvider } from 'react-router-dom';
+import router from './routesConfig';
+
+function AppRouter() {
+  return <RouterProvider router={router} />;
+}
+
+export default AppRouter;
